0.1 Initial Release
Basic game can take a user's guess and will say if it's right or wrong
inside the console. Board functionality is in the works. Most everything
is just temporarily where it's at and will change as time goes on. I
just needed to test some things which is why some stuff is the way it
is.